bin/nvh: Use tar version for platform checks
GNU tar 1.22+ works with `xz` and is found on most Linux distros.
If we have GNU tar 1.22+ and `xz` on PATH, we can extract xz tarballs.

---

bsdtar from libarchive is found on most installs of macOS.
libarchive's bsdtar needs to be version 2.7+,
and needs to be compiled against `liblzma.dylib` on macOS,
in order to extract xz tarballs.

(In Lion and Mountain Lion, libarchive and bsdtar are new enough,
but liblzma.dylib is not present, and bsdtar has xz/lzma support
turned off at compile time.)

On macOS, if `/usr/lib/liblzma.dylib` is present,
and tar is 2.7+, then hopefully tar is compiled against liblzma.

Otherwise, nvh will attempt to use `xz` tarballs,
and fail to extract/install node from them.